+++ Configs/Config.json
@@ -16,18 +16,18 @@
     },
     "debug": {
         "events": {
-            "type": true,
+            "type": false,
             "data": {
                 "READY": false,
                 "MESSAGE_CREATE": false,
-                "INTERACTION_CREATE": true
+                "INTERACTION_CREATE": false
             }
         },
-        "moduleRegister": true,
-        "cache": true,
+        "moduleRegister": false,
+        "cache": false,
         "gateway": false,
         "opCodes": false,
-        "command": true,
-        "init": true
+        "command": false,
+        "init": false
     }
 }
